v.1.0.4 release
mostly fine tuning, the loading log for example only has ONE error in it now and it pertains to acacia blocks.. not sure what the deal is there. otherwise everything is running with every dependancy it could ask for. all smooth, and while lag is never perfect, its sure looking close...